21.7.26

+++ math/quick_pow.cpp
@@ -1,8 +1,8 @@
-int qmi(int a, int k, int p) {
+LL qmi(LL a, int k, int p) {
     LL res = 1;
     while (k) {
-        if (k & 1) res = (LL)res * a % p;
-        a = (LL)a * a % p;
+        if (k & 1) res = res * a % p;
+        a = a * a % p;
         k >>= 1;
     }
     return res;
